change to uint104

+++ contracts/storage/OptionsVaultStorage.sol
@@ -71,7 +71,7 @@ abstract contract OptionsDeltaVaultStorageV1 {
     // % of funds to be used for weekly option purchase
     uint256 public optionAllocationPct;
     // Delta vault equivalent of lockedAmount
-    uint256 public balanceAfterPremium;
+    uint104 public balanceAfterPremium;
 }
 
 // We are following Compound's method of upgrading new contract implementations
